refactored font family
